modify some function prototypes of the dfs_file_ops structure and the function declarations based on it

+++ components/dfs/dfs_v1/filesystems/elmfat/dfs_elm.c
@@ -537,7 +537,7 @@ int dfs_elm_ioctl(struct dfs_file *file, int cmd, void *args)
     return -ENOSYS;
 }
 
-int dfs_elm_read(struct dfs_file *file, void *buf, size_t len)
+ssize_t dfs_elm_read(struct dfs_file *file, void *buf, size_t len)
 {
     FIL *fd;
     FRESULT result;
@@ -560,7 +560,7 @@ int dfs_elm_read(struct dfs_file *file, void *buf, size_t len)
     return elm_result_to_dfs(result);
 }
 
-int dfs_elm_write(struct dfs_file *file, const void *buf, size_t len)
+ssize_t dfs_elm_write(struct dfs_file *file, const void *buf, size_t len)
 {
     FIL *fd;
     FRESULT result;
@@ -596,7 +596,7 @@ int dfs_elm_flush(struct dfs_file *file)
     return elm_result_to_dfs(result);
 }
 
-int dfs_elm_lseek(struct dfs_file *file, off_t offset)
+off_t dfs_elm_lseek(struct dfs_file *file, off_t offset)
 {
     FRESULT result = FR_OK;
     if (file->vnode->type == FT_REGULAR)
